Move ProjectAssetFileWatcher -> NuGet namespace
