+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.VS.UnitTests/ProjectSystem/VS/NuGet/ProjectAssetFileWatcherTests.cs
@@ -4,7 +4,7 @@ using Microsoft.VisualStudio.Shell.Interop;
 using Moq;
 using Xunit;
 
-namespace Microsoft.VisualStudio.ProjectSystem.VS
+namespace Microsoft.VisualStudio.ProjectSystem.VS.NuGet
 {
     [ProjectSystemTrait]
     public class ProjectAssetFileWatcherTests
